soho_menu_harvest/requirements.txt
httpx==0.27.2
aiolimiter==1.1.0
anyio==4.4.0
selectolax==0.3.21
lxml==5.3.0
beautifulsoup4==4.12.3
python-rapidjson==1.14
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.4
rapidfuzz==3.9.3
user-agents==2.2.0
robotexclusionrulesparser==1.7.1
orjson==3.10.7
# optional:
supabase==2.5.3
psycopg2-binary==2.9.9
# optional, guarded by flag:
playwright==1.47.0
